new training implemented

+++ src/data/fields.py
@@ -68,10 +68,13 @@ class PatchPointsField(Field):
         if points.dtype == np.float16:
             points = points.astype(np.float32)
             points += 1e-4 * np.random.randn(*points.shape)
-
+        
+        points = points.astype(np.float32)
+        
         occupancies = points_dict['occupancies']
         if self.unpackbits:
             occupancies = np.unpackbits(occupancies)[:points.shape[0]]
+        
         occupancies = occupancies.astype(np.float32)
 
         # acquire the crop
@@ -135,9 +138,12 @@ class PointsField(Field):
             points = points.astype(np.float32)
             points += 1e-4 * np.random.randn(*points.shape)
 
+        points = points.astype(np.float32)
+        
         occupancies = points_dict['occupancies']
         if self.unpackbits:
             occupancies = np.unpackbits(occupancies)[:points.shape[0]]
+        
         occupancies = occupancies.astype(np.float32)
 
         data = {
